refactor: make product price from transaction data hook
