+++ ui/hooks/subscription/useSubscriptionPricing.ts
@@ -3,14 +3,18 @@ import { useDispatch, useSelector } from 'react-redux';
 import log from 'loglevel';
 import {
   ChainPaymentInfo,
+  PAYMENT_TYPES,
   PaymentType,
   PricingPaymentMethod,
   PricingResponse,
+  PRODUCT_TYPES,
   ProductPrice,
   ProductType,
+  RECURRING_INTERVALS,
   TokenPaymentInfo,
 } from '@metamask/subscription-controller';
 import { Hex } from '@metamask/utils';
+import { TransactionMeta } from '@metamask/transaction-controller';
 import { getSubscriptionPricing } from '../../selectors/subscription';
 import {
   getSubscriptionCryptoApprovalAmount,
@@ -25,6 +29,7 @@ import {
   NativeAsset,
 } from '../../components/multichain/asset-picker-amount/asset-picker-modal/types';
 import { AssetType } from '../../../shared/constants/transaction';
+import { useAsyncResult } from '../useAsync';
 
 export type TokenWithApprovalAmount = (
   | AssetWithDisplayData<ERC20Asset>
@@ -208,4 +213,86 @@ export const useSubscriptionPaymentMethods = (
       ),
     [pricing, paymentType],
   );
+};
+
+/**
+ * Use this hook to get the shield subscription price derived from transaction data.
+ *
+ * @param params - The parameters for the hook.
+ * @param params.transactionMeta - The transaction meta.
+ * @param params.decodedApprovalAmount - The decoded approval amount.
+ * @returns The product price.
+ */
+export const useShieldSubscriptionPricingFromTokenApproval = ({
+  transactionMeta,
+  decodedApprovalAmount,
+}: {
+  transactionMeta?: TransactionMeta;
+  decodedApprovalAmount?: string;
+}) => {
+  const { subscriptionPricing } = useSubscriptionPricing();
+  const pricingPlans = useSubscriptionProductPlans(
+    PRODUCT_TYPES.SHIELD,
+    subscriptionPricing,
+  );
+  const cryptoPaymentMethod = useSubscriptionPaymentMethods(
+    PAYMENT_TYPES.byCrypto,
+    subscriptionPricing,
+  );
+  const selectedTokenPrice = useMemo(() => {
+    return cryptoPaymentMethod?.chains
+      ?.find(
+        (chain) =>
+          chain.chainId.toLowerCase() ===
+          transactionMeta?.chainId.toLowerCase(),
+      )
+      ?.tokens.find(
+        (token) =>
+          token.address.toLowerCase() ===
+          transactionMeta?.txParams?.to?.toLowerCase(),
+      );
+  }, [cryptoPaymentMethod, transactionMeta]);
+
+  // need to do async here since `getSubscriptionCryptoApprovalAmount` make call to background script
+  const { value: productPrice } = useAsyncResult(async (): Promise<
+    ProductPrice | undefined
+  > => {
+    if (selectedTokenPrice) {
+      const params = {
+        chainId: transactionMeta?.chainId as Hex,
+        paymentTokenAddress: selectedTokenPrice.address as Hex,
+        productType: PRODUCT_TYPES.SHIELD,
+      };
+      const [monthlyApprovalAmount, yearlyApprovalAmount] = await Promise.all([
+        getSubscriptionCryptoApprovalAmount({
+          ...params,
+          interval: RECURRING_INTERVALS.month,
+        }),
+        getSubscriptionCryptoApprovalAmount({
+          ...params,
+          interval: RECURRING_INTERVALS.year,
+        }),
+      ]);
+
+      if (monthlyApprovalAmount.approveAmount === decodedApprovalAmount) {
+        return pricingPlans?.find(
+          (plan) => plan.interval === RECURRING_INTERVALS.month,
+        );
+      }
+      if (yearlyApprovalAmount.approveAmount === decodedApprovalAmount) {
+        return pricingPlans?.find(
+          (plan) => plan.interval === RECURRING_INTERVALS.year,
+        );
+      }
+    }
+
+    return undefined;
+  }, [
+    transactionMeta,
+    selectedTokenPrice,
+    decodedApprovalAmount,
+    pricingPlans,
+  ]);
+
+  return { productPrice };
 };
