+++ ui/pages/confirmations/components/confirm/info/shield-subscription-approve/shield-subscription-approve.tsx
@@ -1,26 +1,15 @@
 import { Box } from '@metamask/design-system-react';
-import {
-  PAYMENT_TYPES,
-  PRODUCT_TYPES,
-  ProductPrice,
-  RECURRING_INTERVALS,
-} from '@metamask/subscription-controller';
+import { PRODUCT_TYPES } from '@metamask/subscription-controller';
 import { TransactionMeta } from '@metamask/transaction-controller';
 import { Hex } from '@metamask/utils';
 import { BigNumber } from 'bignumber.js';
-import React, { useMemo } from 'react';
+import React from 'react';
 import { useUserSubscriptions } from '../../../../../../hooks/subscription/useSubscription';
 import { useConfirmContext } from '../../../../context/confirm';
 import { useAssetDetails } from '../../../../hooks/useAssetDetails';
 import { useDecodedTransactionData } from '../hooks/useDecodedTransactionData';
 import { GasFeesSection } from '../shared/gas-fees-section/gas-fees-section';
-import {
-  useSubscriptionPaymentMethods,
-  useSubscriptionPricing,
-  useSubscriptionProductPlans,
-} from '../../../../../../hooks/subscription/useSubscriptionPricing';
-import { useAsyncResult } from '../../../../../../hooks/useAsync';
-import { getSubscriptionCryptoApprovalAmount } from '../../../../../../store/actions';
+import { useShieldSubscriptionPricingFromTokenApproval } from '../../../../../../hooks/subscription/useSubscriptionPricing';
 import { AccountDetails } from './account-details';
 import { EstimatedChanges } from './estimated-changes';
 import ShieldSubscriptionApproveLoader from './shield-subscription-approve-loader';
@@ -47,69 +36,10 @@ const ShieldSubscriptionApproveInfo = () => {
     .div(10 ** (decimals ?? 0))
     .toFixed();
 
-  const { subscriptionPricing } = useSubscriptionPricing();
-  const pricingPlans = useSubscriptionProductPlans(
-    PRODUCT_TYPES.SHIELD,
-    subscriptionPricing,
-  );
-  const cryptoPaymentMethod = useSubscriptionPaymentMethods(
-    PAYMENT_TYPES.byCrypto,
-    subscriptionPricing,
-  );
-  const selectedTokenPrice = useMemo(() => {
-    return cryptoPaymentMethod?.chains
-      ?.find(
-        (chain) =>
-          chain.chainId.toLowerCase() ===
-          transactionMeta?.chainId.toLowerCase(),
-      )
-      ?.tokens.find(
-        (token) =>
-          token.address.toLowerCase() ===
-          transactionMeta?.txParams?.to?.toLowerCase(),
-      );
-  }, [cryptoPaymentMethod, transactionMeta]);
-
-  // need to do async here since `getSubscriptionCryptoApprovalAmount` make call to background script
-  const { value: productPrice } = useAsyncResult(async (): Promise<
-    ProductPrice | undefined
-  > => {
-    if (selectedTokenPrice) {
-      const params = {
-        chainId: transactionMeta?.chainId as Hex,
-        paymentTokenAddress: selectedTokenPrice.address as Hex,
-        productType: PRODUCT_TYPES.SHIELD,
-      };
-      const [monthlyApprovalAmount, yearlyApprovalAmount] = await Promise.all([
-        getSubscriptionCryptoApprovalAmount({
-          ...params,
-          interval: RECURRING_INTERVALS.month,
-        }),
-        getSubscriptionCryptoApprovalAmount({
-          ...params,
-          interval: RECURRING_INTERVALS.year,
-        }),
-      ]);
-
-      if (monthlyApprovalAmount.approveAmount === decodedApprovalAmount) {
-        return pricingPlans?.find(
-          (plan) => plan.interval === RECURRING_INTERVALS.month,
-        );
-      }
-      if (yearlyApprovalAmount.approveAmount === decodedApprovalAmount) {
-        return pricingPlans?.find(
-          (plan) => plan.interval === RECURRING_INTERVALS.year,
-        );
-      }
-    }
-
-    return undefined;
-  }, [
+  const { productPrice } = useShieldSubscriptionPricingFromTokenApproval({
     transactionMeta,
-    selectedTokenPrice,
     decodedApprovalAmount,
-    pricingPlans,
-  ]);
+  });
 
   const { trialedProducts, loading: subscriptionsLoading } =
     useUserSubscriptions();
